fix(evpn): netlink split changes
